Decompile sub_38A0

+++ include/functions.h
@@ -105,7 +105,7 @@ extern void sub_374C(void);
 extern void sub_377C(void);
 extern void sub_37B4(s8);
 extern void sub_3828(s8, s8);
-extern void sub_38A0(u32, u32);
+extern void sub_38A0(s8, u16);
 extern s16 GetEReaderCardIndex(void);
 extern void sub_394C(void);
 extern void sub_3AB4(void);
